✨ feat: add job to delete all undefined playlists (#46)
* ✨ feat: Add endpoint to clear 404 playlists and update database

This commit introduces a new endpoint in the server file `delete-undefined-playlists/+server.ts` that handles the clearing of 404 playlists. It includes:
- Import statements for necessary modules
- Definition of a GET RequestHandler to clear 404 playlists based on a key
- Invocation of `clear404Playlists` method in `ytMusicController` to handle actual deletion logic
- Logging information about the number of playlists deleted

Additionally, this commit incorporates changes in `YtMusicController.ts`, `PlaylistsRepository.ts`, and `YTMusicService.ts` to support the clearing of 404 playlists and updating related data in the database. It also adds a Python script `clear_404_playlists.py` under `static/python-scripts` to facilitate the identification of deleted playlists within the YTMusic service.

* feat: reconfigure modals on frontend

* feat: add backend playlist deletion logic